requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
black==22.1.0
blue==0.9.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.3
django-celery-results==2.5.1
django-ninja==1.3.0
flake8==4.0.1
idna==3.10
kombu==5.4.2
mccabe==0.6.1
mypy-extensions==1.0.0
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pycodestyle==2.8.0
pydantic==2.10.1
pydantic_core==2.27.1
pyflakes==2.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.2.0
requests==2.32.3
scalar-django-ninja @ git+https://github.com/scalar/scalar.git@e3599563693ce80c5b9cbdb9caf348856a67522d#subdirectory=packages/scalar_django_ninja
six==1.16.0
sqlparse==0.5.2
tomli==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
